dflash: uplift DFlash runtime from qwen35/ to common/
Refactor the DFlash runtime so all generic infrastructure lives in
src/common/ and qwen35/ retains only model-specific code. This unblocks
sharing the spec-decode loop, feature transfer, and IPC machinery with
gemma4, laguna, and any future target architecture.

New shared infrastructure (src/common/):
  - dflash_feature_ring.{h,cpp}     DraftFeatureMirror + ring copy helpers
                                    (decoupled from qwen35 TargetCache;
                                    sync_range/sync_tail now take generic
                                    src_target_feat / src_cap pair)
  - dflash_capture.{h,cpp}          target_capture_index(ids, n, layer)
                                    (decoupled from TargetWeights)
  - dflash_draft_ipc.{h,cpp}        DFlashDraftIpcClient + remote feature
                                    copy + stream_status; daemon entry
                                    declared here, body lives in qwen35
  - dflash_layer_split_runtime.h    LayerSplitRuntimeConfig + ActivationPair
  - dflash_spec_decode.{h,cpp}      Generic run_dflash_spec_decode loop
                                    typed against DFlashTarget — drives any
                                    target through verify_batch /
                                    snapshot_kv / restore_kv / embed_tokens
                                    / project_hidden_to_tokens / is_eos

New qwen35 adapters:
  - qwen35_layer_split_dflash_target.{h,cpp}  Multi-GPU DFlashTarget
        adapter over std::vector<TargetLayerSplitShard>. Holds feature
        ring, kq_stride_pad, fa_window, optional remote-draft IPC
        internally. verify_batch delegates to
        run_target_layer_split_forward; snapshot/restore iterate every
        shard; project_hidden_to_tokens builds the LM-head graph on
        shards.back().
  - draft_ipc_daemon.cpp            qwen35-side body of the DFlash IPC
        daemon (uses qwen-specific DraftWeights / build_draft_step).

Removed (replaced by common/ + adapter):
  - qwen35/feature_copy.{h,cpp}
  - qwen35/draft_feature_mirror.{h,cpp}
  - qwen35/draft_ipc.{h,cpp}
  - qwen35/spec_decode.{h,cpp}

Modified:
  - qwen35/layer_split_types.h trimmed to TargetLayerSplitShard +
    find_target_shard; LayerSplitRuntimeConfig/ActivationPair moved to
    common.
  - qwen35/layer_split_daemon.{h,cpp} now constructs
    Qwen35LayerSplitDFlashTarget and calls run_dflash_spec_decode.
  - test/test_dflash.cpp same call-site update.
  - 13 other call sites + CMakeLists wired to the new APIs.
  - docs/ARCHITECTURE.md updated component tree + spec-decode section.

Why no DFlashTarget extension was needed: every spec_decode operation
already had a counterpart on the existing interface. The only
model-specific bits (kq_stride_pad, fa_window, multi-shard fan-out, IPC
remote draft) are encapsulated behind the adapter constructor. Local
draft hidden states route through host (~160 KB per step) before
LM-head projection so the loop never touches target-internal device
layout — matches the IPC remote-draft path.

Out of scope: layer_split_daemon_loop.cpp stays in qwen35/ — it loads
qwen35 GGUF and constructs qwen35 caches/shards, so it is qwen35
daemon glue, not generic runtime infra.

Verification: full tree builds clean (cmake --build . → all 27 targets
including dflash27b, test_dflash, smoke/bench binaries). Runtime
regression decode not run in this change; recommend exercising the
qwen35 layer-split smoke or test_dflash with a real prompt before
merging to confirm bit-equivalent behavior.
